refactor(routing): date-parser 를 엔진/패턴 테이블로 분리 — 200줄 경계 준수
시간 표현이 늘며 231줄이 됐다. 패턴 테이블(date-patterns.ts)과 공통 타입(date-types.ts)을
떼어내 엔진은 58줄. 라우팅 결과 불변(릭 76행 도달도구·파라미터 완전 일치).
