+++ src/lib/date-types.ts
@@ -1,0 +1,8 @@
+/** 날짜 파서 공통 타입 */
+
+export interface DateRange {
+  /** 시작일 (YYYYMMDD) */
+  from: string
+  /** 종료일 (YYYYMMDD) */
+  to: string
+}
